api/requirements/base.txt
# python==3.8

# Tools
pre-commit

# Django
Django==3.2
django-environ==0.7.0
six==1.16.0

# Python-PostgreSQL Database Adapter
psycopg2-binary==2.9.1

# Models
django-model-utils==4.1.1
python_dateutil==2.8.2

# Utils
django-ipware==4.0.0

# DRF
djangorestframework==3.12.4
django-oauth-toolkit==1.5.0
django-cors-headers==3.9.0
django-rest-auth==0.9.5
django-allauth==0.45.0
django_filter==21.1

# Tagging
django-taggit==1.5.1

# Sendgird
sendgrid-django==4.2.0

# Static files
# whitenoise==4.1.4  # https://github.com/evansd/whitenoise

# Document
drf-yasg==1.20.0

# Celery
celery==5.2.0
redis==3.5.3
django-celery-beat==2.2.1
django-celery-results==2.2.0
celery-once==3.0.1
eventlet==0.32.0

# debug
django_extensions==3.1.3
django-debug-toolbar==3.2.2

# WSGI Handler
gunicorn==20.1.0

# code coverage and testing
django-webtest==1.9.7
coverage==6.0

# storage
boto3==1.18.53
django-storages==1.11.1
Pillow==8.3.2

# user-agent
pyyaml==5.4.1
ua-parser==0.10.0
user-agents==2.2.0
django-user-agents==0.4.0
phonenumbers==8.12.33

# Faker
Faker==8.14.1

# Socket IO emitter
socket.io-emitter==0.1.5.1

# JWT
djangorestframework_simplejwt==5.0.0

# Bug Service
sentry-sdk==1.4.3
bugsnag==4.1.0

# (Mime type) Upload app
python_magic==0.4.24

# logstash
python-logstash==0.4.6

# tsvector-field
django-tsvector-field==0.9.5

# 2FA
django-otp==1.1.1

# bdist_wheel
wheel==0.37.0

# Django cache memoize
django-cache-memoize==0.1.10

# django health check plus
django-health-check-plus==2.0.1

# Twython
twython==3.9.1

# location
us==2.0.2

#branch io
branchio==0.1.1

# calculator workdays
workdays==1.4

# django role permissions
django-role-permissions==3.1.1

# FCM
fcm-django==1.0.5

# django-property-filter
django-property-filter==1.1.0
